Do not hoist jest-image-snapshot (and its types)
Tell Yarn not to hoist the `jest-image-snapshot` package (and the types for it). The default hoisting behavior was causing it to show up in the type definitions that get built for `jest-mock-web-client`, even though it does not use `jest-image-snapshot`, which in turn cause errors using it in other packages that don't need `jest-image-snapshot`.

This should fix the first part of #118.