Generate dynamic mock of Slack `WebClient`
This process is a little complex, but it will generate a mock that matches the interface of whatever version of Slack is currently installed, meaning this module does not need to manually add or remove properties/methods as they change in the underlying `slack/web-api` package, and should work seamlessly with future versions.

This should also make `MockedWebClient` constructable, and fixes #104.